LayoutTests/fast/forms/range/disabled-while-dragging-expected.txt

Tests that click events still work after a range is disabled while dragging.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS range.disabled is false
Drag range slider.
PASS range.disabled is true
Click button
PASS buttonClicked is true
PASS successfullyParsed is true

TEST COMPLETE
 Click
